Adds to gameboard and adds Player factory
Added functionality for the board to keep track of missed attacks and
can also return if any ships are left on the board or not. Added a
Player factory with the ability to attack a coordinate on a board,
check if he/she has already attacked a certain square, and the ability
to attack a random square. Added tests for the player as well