Adds UI elements for game interaction
Added functions to render the gameboards. Can click on a square to
attack that position. The board shows if the attack was a hit, miss, or
if you've sunk the ship. After you attack, the CPU will attack a random
position on your board. CPU randomly places ships on the board. Modal
pops up if all ships are sunk on a board and displays the winner.